fix tick count

+++ src/robocars_autopilot.cpp
@@ -105,7 +105,7 @@ class onRunningMode
         void react(TickEvent const & e) override {
             RobocarsStateMachine::react(e);
             __tick_count++;
-            if (__tick_count%(2000/loop_hz)==0) {
+            if ((__tick_count%(uint32_t)(2000/loop_hz))==0) {
                 ri->reportStats();
                 ri->updateParam();
             }
